refactor: BigDecimal로 베팅 금액 계산

+++ src/main/java/view/responsedto/DealerResultDto.java
@@ -1,5 +1,7 @@
 package view.responsedto;
 
-public record DealerResultDto(int dealerProfit) {
+import java.math.BigDecimal;
+
+public record DealerResultDto(BigDecimal dealerProfit) {
 
 }
